fixed issue, that new message gets deleted

+++ inu/ext/commands/music.py
@@ -1292,9 +1292,10 @@ async def queue(
             components=await build_music_components(node=node), 
             **kwargs
         )
-        music_messages[ctx.guild_id] = await msg.message()
+        new_music_msg = await msg.message()
+        music_messages[ctx.guild_id] = new_music_msg
         try:
-            if not old_music_msg is None:
+            if not old_music_msg is None and not old_music_msg.id == new_music_msg.id:
                 await old_music_msg.delete()
         except hikari.NotFoundError:
             pass
